requirements.txt
flask
gunicorn
flask_sqlalchemy
python-docx
PyPDF2
